+++ deltafi-core/src/test/java/org/deltafi/core/datafetchers/DeltaFilesDatafetcherTestHelper.java
@@ -18,6 +18,7 @@
 package org.deltafi.core.datafetchers;
 
 import org.deltafi.common.content.ContentReference;
+import org.deltafi.common.content.Segment;
 import org.deltafi.common.types.Content;
 import org.deltafi.common.types.IngressInput;
 import org.deltafi.common.types.KeyValue;
@@ -44,9 +45,9 @@ public class DeltaFilesDatafetcherTestHelper {
     static final String DID_2 = UUID.randomUUID().toString();
     static final List<KeyValue> METADATA = Arrays.asList(new KeyValue("k1", "v1"), new KeyValue("k2", "v2"));
     static final String MEDIA_TYPE = "plain/text";
-    static final ContentReference CONTENT_REFERENCE = new ContentReference(OBJECT_UUID, 0, SIZE, DID, MEDIA_TYPE);
+    static final ContentReference CONTENT_REFERENCE = new ContentReference(MEDIA_TYPE, new Segment(OBJECT_UUID, 0, SIZE, DID));
     static final List<Content> CONTENT = Collections.singletonList(Content.newBuilder().contentReference(CONTENT_REFERENCE).build());
-    static final ContentReference CONTENT_REFERENCE_2 = new ContentReference(OBJECT_UUID_2, 0, SIZE, DID, MEDIA_TYPE);
+    static final ContentReference CONTENT_REFERENCE_2 = new ContentReference(MEDIA_TYPE, new Segment(OBJECT_UUID_2, 0, SIZE, DID));
     static final List<Content> CONTENT_2 = Collections.singletonList(Content.newBuilder().contentReference(CONTENT_REFERENCE_2).build());
     static final SourceInfo SOURCE_INFO = new SourceInfo(FILENAME, FLOW, METADATA);
     static final SourceInfo SOURCE_INFO_EMPTY_METADATA = new SourceInfo(FILENAME, FLOW, Collections.emptyList());
@@ -80,11 +81,14 @@ public class DeltaFilesDatafetcherTestHelper {
                 .value()
                 .parent()
                 .contentReference()
+                .mediaType()
+                .size()
+                .segments()
                 .uuid()
                 .offset()
                 .size()
                 .did()
-                .mediaType()
+                .parent()
                 .parent()
                 .metadata()
                 .key()
@@ -114,10 +118,14 @@ public class DeltaFilesDatafetcherTestHelper {
                 .filename()
                 .formatAction()
                 .contentReference()
+                .mediaType()
+                .size()
+                .segments()
                 .uuid()
                 .offset()
                 .size()
                 .did()
+                .parent()
                 .parent()
                 .parent()
                 .filtered()
@@ -155,11 +163,14 @@ public class DeltaFilesDatafetcherTestHelper {
                   .value()
                   .parent()
                 .contentReference()
-                  .uuid()
-                  .offset()
-                  .size()
-                  .did()
                   .mediaType()
+                  .size()
+                  .segments()
+                    .uuid()
+                    .offset()
+                    .size()
+                    .did()
+                    .parent()
                   .parent()
                 .parent()
               .metadata()
@@ -189,10 +200,14 @@ public class DeltaFilesDatafetcherTestHelper {
                 .filename()
                 .formatAction()
                 .contentReference()
-                    .uuid()
-                    .offset()
+                    .mediaType()
                     .size()
-                    .did()
+                    .segments()
+                        .uuid()
+                        .offset()
+                        .size()
+                        .did()
+                        .parent()
                     .parent()
                 .parent()
             .egressed()