+++ deltafi-core/src/test/java/org/deltafi/core/services/DeltaFilesServiceTest.java
@@ -23,6 +23,7 @@ import org.assertj.core.api.Assertions;
 import org.deltafi.common.action.ActionEventQueue;
 import org.deltafi.common.content.ContentReference;
 import org.deltafi.common.content.ContentStorageService;
+import org.deltafi.common.content.Segment;
 import org.deltafi.common.metrics.MetricRepository;
 import org.deltafi.common.metrics.MetricsUtil;
 import org.deltafi.common.types.*;
@@ -61,7 +62,7 @@ class DeltaFilesServiceTest {
     ContentStorageService contentStorageService;
 
     @Captor
-    ArgumentCaptor<List<ContentReference>> contentReferenceCaptor;
+    ArgumentCaptor<List<Segment>> segmentCaptor;
 
     @Captor
     ArgumentCaptor<List<DeltaFile>> deltaFileListCaptor;
@@ -94,7 +95,7 @@ class DeltaFilesServiceTest {
 
         String did = UUID.randomUUID().toString();
 
-        List<Content> content = Collections.singletonList(Content.newBuilder().contentReference(new ContentReference()).build());
+        List<Content> content = Collections.singletonList(Content.newBuilder().contentReference(new ContentReference("mediaType")).build());
         IngressInput ingressInput = new IngressInput(did, sourceInfo, content, OffsetDateTime.now());
 
         DeltaFile deltaFile = deltaFilesService.ingress(ingressInput);
@@ -108,7 +109,7 @@ class DeltaFilesServiceTest {
     @Test
     void setThrowsOnMissingFlow() {
         SourceInfo sourceInfo = new SourceInfo(null, "nonsense", List.of());
-        List<Content> content = Collections.singletonList(Content.newBuilder().contentReference(new ContentReference()).build());
+        List<Content> content = Collections.singletonList(Content.newBuilder().contentReference(new ContentReference("mediaType")).build());
         IngressInput ingressInput = new IngressInput("did", sourceInfo, content, OffsetDateTime.now());
 
         when(flowService.getRunningFlowByName(sourceInfo.getFlow())).thenThrow(new DgsEntityNotFoundException());
@@ -214,17 +215,17 @@ class DeltaFilesServiceTest {
     @Test
     void testDelete() {
         DeltaFile deltaFile1 = Util.buildDeltaFile("1");
-        ContentReference cr1 = new ContentReference("a", "1", "mediaType");
+        ContentReference cr1 = new ContentReference("mediaType", new Segment("a", "1"));
         deltaFile1.getProtocolStack().add(ProtocolLayer.builder().content(List.of(Content.newBuilder().contentReference(cr1).build())).build());
-        ContentReference cr2 = new ContentReference("b", "2", "mediaType");
+        ContentReference cr2 = new ContentReference("mediaType", new Segment("b", "2"));
         DeltaFile deltaFile2 = Util.buildDeltaFile("2");
         deltaFile2.getFormattedData().add(FormattedData.newBuilder().contentReference(cr2).build());
         when(deltaFileRepo.findForDelete(any(), any(), anyLong(), any(), any(), anyBoolean(), anyInt())).thenReturn(List.of(deltaFile1, deltaFile2));
 
         deltaFilesService.delete(OffsetDateTime.now().plusSeconds(1), null, 0L, null, "policy", false, 10);
 
-        verify(contentStorageService).deleteAll(contentReferenceCaptor.capture());
-        assertEquals(List.of(cr1, cr2), contentReferenceCaptor.getValue());
+        verify(contentStorageService).deleteAll(segmentCaptor.capture());
+        assertEquals(List.of(cr1.getSegments().get(0), cr2.getSegments().get(0)), segmentCaptor.getValue());
         verify(deltaFileRepo).saveAll(deltaFileListCaptor.capture());
         assertEquals(List.of(deltaFile1, deltaFile2), deltaFileListCaptor.getValue());
         assertNotNull(deltaFile1.getContentDeleted());
@@ -235,17 +236,17 @@ class DeltaFilesServiceTest {
     @Test
     void testDeleteMetadata() {
         DeltaFile deltaFile1 = Util.buildDeltaFile("1");
-        ContentReference cr1 = new ContentReference("a", "1", "mediaType");
+        ContentReference cr1 = new ContentReference("mediaType", new Segment("a", "1"));
         deltaFile1.getProtocolStack().add(ProtocolLayer.builder().content(List.of(Content.newBuilder().contentReference(cr1).build())).build());
-        ContentReference cr2 = new ContentReference("b", "2", "mediaType");
+        ContentReference cr2 = new ContentReference("mediaType", new Segment("b", "2"));
         DeltaFile deltaFile2 = Util.buildDeltaFile("2");
         deltaFile2.getFormattedData().add(FormattedData.newBuilder().contentReference(cr2).build());
         when(deltaFileRepo.findForDelete(any(), any(), anyLong(), any(), any(), anyBoolean(), anyInt())).thenReturn(List.of(deltaFile1, deltaFile2));
 
         deltaFilesService.delete(OffsetDateTime.now().plusSeconds(1), null, 0L, null, "policy", true, 10);
 
-        verify(contentStorageService).deleteAll(contentReferenceCaptor.capture());
-        assertEquals(List.of(cr1, cr2), contentReferenceCaptor.getValue());
+        verify(contentStorageService).deleteAll(segmentCaptor.capture());
+        assertEquals(List.of(cr1.getSegments().get(0), cr2.getSegments().get(0)), segmentCaptor.getValue());
         verify(deltaFileRepo, never()).saveAll(any());
         verify(deltaFileRepo).deleteAll(deltaFileListCaptor.capture());
         assertEquals(List.of(deltaFile1, deltaFile2), deltaFileListCaptor.getValue());
@@ -291,11 +292,11 @@ class DeltaFilesServiceTest {
 
     @Test
     void testCalculateBytes() {
-        ContentReference contentReference1 = new ContentReference("uuid1", 0, 500, "did1", "*/*");
-        ContentReference contentReference2 = new ContentReference("uuid1", 400, 200, "did1", "*/*");
-        ContentReference contentReference3 = new ContentReference("uuid1", 200, 200, "did1", "*/*");
-        ContentReference contentReference4 = new ContentReference("uuid2", 5, 200, "did1", "*/*");
-        ContentReference contentReference5 = new ContentReference("uuid3", 5, 200, "did2", "*/*");
+        ContentReference contentReference1 = new ContentReference("*/*", new Segment("uuid1", 0, 500, "did1"));
+        ContentReference contentReference2 = new ContentReference("*/*", new Segment("uuid1", 400, 200, "did1"));
+        ContentReference contentReference3 = new ContentReference("*/*", new Segment("uuid1", 200, 200, "did1"));
+        ContentReference contentReference4 = new ContentReference("*/*", new Segment("uuid2", 5, 200, "did1"));
+        ContentReference contentReference5 = new ContentReference("*/*", new Segment("uuid3", 5, 200, "did2"));
 
         DeltaFile deltaFile = DeltaFile.newBuilder()
                 .protocolStack(List.of(
@@ -390,14 +391,6 @@ class DeltaFilesServiceTest {
     }
 
     private ContentReference createContentReference(String did) {
-        ContentReference contentReference = new ContentReference();
-
-        contentReference.setSize(32L);
-        contentReference.setDid(did);
-        contentReference.setUuid(UUID.randomUUID().toString());
-        contentReference.setOffset(0L);
-        contentReference.setMediaType(APPLICATION_XML);
-
-        return contentReference;
+        return new ContentReference(APPLICATION_XML, new Segment(UUID.randomUUID().toString(), 0L, 32L, did));
     }
 }